MSVC fixes for GDI printing
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1284 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/wc3270/gdi_print.c
@@ -454,8 +454,8 @@ gdi_init(const char *printer_name, const char **fail)
 		maxphmargin, maxpvmargin);
 
 	/* Compute the margins in pixels. */
-	pstate.hmargin_pixels = uparm.hmargin * pchar.ppiX;
-	pstate.vmargin_pixels = uparm.vmargin * pchar.ppiY;
+	pstate.hmargin_pixels = (int)(uparm.hmargin * pchar.ppiX);
+	pstate.vmargin_pixels = (int)(uparm.vmargin * pchar.ppiY);
 
 	/* See if the margins are too small. */
 	if (pstate.hmargin_pixels < maxphmargin) {
@@ -652,7 +652,7 @@ gdi_screenful(struct ea *ea, unsigned short rows, unsigned short cols,
 	if (pstate.out_row == 0 && pstate.caption != NULL) {
 		int center;
 
-		if (strlen(pstate.caption) < pstate.usable_cols) {
+		if ((int)strlen(pstate.caption) < pstate.usable_cols) {
 			center = (pstate.usable_xpixels -
 			       (strlen(pstate.caption) * pstate.space_size.cx))
 			      / 2;
